refactor: split ai_engine.py (2016 LOC) into subpackage per AGENTS.md Rule 14
- Create mcp_video/ai_engine/ subpackage with 9 modules:
  transcribe.py (211) - ai_transcribe, _format_srt/txt/md/json
  spatial.py (281) - audio_spatial, _apply_simple_spatial, _azimuth_to_pan, _elevation_to_volume
  scene.py (136) - ai_scene_detect, _standard_scene_detect re-export
  silence.py (278) - ai_remove_silence, silence detection helpers
  stem.py (139) - ai_stem_separation
  color.py (220) - ai_color_grade, _match_reference_colors
  upscale.py (448) - ai_upscale, _ai_upscale_opencv, _verify_model_hash, _MODEL_HASHES
  download.py (258) - URL/download helpers (_resolve_video_source, _is_safe_url, etc.)
  __init__.py (265) - analyze_video + re-exports + test monkeypatch compat
- Preserves backward-compatible import path and test monkeypatch targets

Lint: all checks passed
Tests: 773 passed, 9 skipped, 2 xpassed
